Rename Payara module to include base version
