+++ instrumentation/payara-5.2020/javaagent/build.gradle.kts
@@ -5,5 +5,5 @@ plugins {
 // No muzzle check because this instrumentation is written in ASM and muzzle won't work with it.
 
 dependencies {
-  library("fish.payara.extras:payara-embedded-web:5.2021.2")
+  library("fish.payara.extras:payara-embedded-web:5.2020.2")
 }
