+++ instrumentation/payara-5.2020/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/payara/v5_2020/StandardWrapperInstrumentation.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.payara;
+package io.opentelemetry.javaagent.instrumentation.payara.v5_2020;
 
 import static net.bytebuddy.matcher.ElementMatchers.named;
