+++ settings.gradle.kts
@@ -540,7 +540,7 @@ include(":instrumentation:oracle-ucp-11.2:testing")
 include(":instrumentation:oshi:javaagent")
 include(":instrumentation:oshi:library")
 include(":instrumentation:oshi:testing")
-include(":instrumentation:payara:javaagent")
+include(":instrumentation:payara-5.2020:javaagent")
 include(":instrumentation:pekko:pekko-actor-1.0:javaagent")
 include(":instrumentation:pekko:pekko-http-1.0:javaagent")
 include(":instrumentation:play:play-mvc:play-mvc-2.4:javaagent")
